Directory tree and file naming fix
