+++ src/store/profile/index.js
@@ -1,4 +1,4 @@
-import {useDispatch, useSelector} from "react-redux";
+import {useDispatch} from "react-redux";
 import {useCallback, useState} from "react";
 import {changeName, toggleShowName} from "./actions";
